working ldos maker.

+++ runscripthandler.py
@@ -84,7 +84,8 @@ class RunScriptHandler():
         self.PDOS_DeltaE     = kwargs.get("PDOS_DeltaE",0.1)
         self.PDOS_degauss    = kwargs.get("PDOS_degauss",0.01)
         self.PDOS_required_projections = kwargs.get("PDOS_required_projections", [["Hg","all"],["Hg","s"],["Hg","p"],["Hg","d"],["Te","all"],["Te","s"],["Te","p"],["Te","d"],["H","s"],["S","all"],["S","s"],["S","p"],["C","all"],["C","s"],["C","p"]])
-        self.n_proj_boxes    = kwargs.get("n_proj_boxes", 32)
+        self.fft_grid        = kwargs.get("fft_grid", [40, 40, 432])
+        self.n_proj_boxes    = kwargs.get("n_proj_boxes", self.fft_grid[2])
 
         # Quantum espresso inits some other inits that need to be only set if explicitly given can be found below this.
         self.espresso_inputs = {"pseudopotentials": self.pseudopotentials,
@@ -271,31 +272,24 @@ class RunScriptHandler():
                 file.write(f"  degauss = {self.PDOS_degauss}\n")
                 file.write(f"/\n")
             os.rename(f"{self.identifier}.pdos.in", f"./{self.base_folder}/{run_name}/{self.identifier}.pdos.in")  
-        # else:
-        #     ase.io.write(f"{self.identifier}.scf.in", self.atoms_object, format = "espresso-in", **self.espresso_inputs)
-        #     os.rename(f"{self.identifier}.scf.in", f"./{self.base_folder}/{run_name}/{self.identifier}.scf.in")
 
         if "ldos" in self.calculation:
             # We are not using ASE to write this file. It is simple and that is one reason
             with open(f"{self.identifier}.ldos.in", "w+") as file:
                 file.write(f"&projwfc\n")
-                file.write(f"  prefix       = '{self.espresso_inputs['prefix']}'\n")
+                file.write(f"  prefix       = '{self.espresso_inputs['prefix']}',\n")
                 # file.write(f"  filpdos = '{self.espresso_inputs['prefix']}'\n")
                 # file.write(f"  Emin    = {self.PDOS_EMIN}\n")
                 # file.write(f"  Emax    = {self.PDOS_EMAX}\n")
-                file.write(f"  DeltaE       = {self.PDOS_DeltaE}\n")
-                file.write(f"  degauss      = {self.PDOS_degauss}\n")
+                file.write(f"  DeltaE       = {self.PDOS_DeltaE},\n")
+                file.write(f"  degauss      = {self.PDOS_degauss},\n")
                 file.write(f"  tdosinboxes  = .true.,\n")
                 file.write(f"  plotboxes    = .true.,\n")
                 file.write(f"  n_proj_boxes = {self.n_proj_boxes},\n")
-                file.write(f"/\n")
-
                 for i in range(1, self.n_proj_boxes+1):
-                    file.write(f"irmin(1,{i})=1, irmax(1,{i})={self.n_proj_boxes}, irmin(2,{i})=1, irmax(2,{i})={self.n_proj_boxes}, irmin(3,{i})=1, irmax(3,{i})={i},\n")
+                    file.write(f"  irmin(1,{i})=1, irmax(1,{i})={self.fft_grid[0]}, irmin(2,{i})=1, irmax(2,{i})={self.fft_grid[1]}, irmin(3,{i})={i}, irmax(3,{i})={i},\n")
+                file.write(f"/\n")
             os.rename(f"{self.identifier}.ldos.in", f"./{self.base_folder}/{run_name}/{self.identifier}.ldos.in")  
-        # else:
-        #     ase.io.write(f"{self.identifier}.scf.in", self.atoms_object, format = "espresso-in", **self.espresso_inputs)
-        #     os.rename(f"{self.identifier}.scf.in", f"./{self.base_folder}/{run_name}/{self.identifier}.scf.in")
 
 
 
@@ -543,8 +537,9 @@ class RunScriptHandler():
             file.write(f"#SBATCH --ntasks={self.ntasks} # Number of taks per node for the job\n")
             file.write(f"#SBATCH --mail-user=erathnayake@sivananthanlabs.us\n")
             file.write(f"#SBATCH --mail-type=ALL\n")
-            file.write(f"mpirun -np {self.ntasks} pw.x -npools {self.npools} < {self.identifier}.scf.in > {self.identifier}.scf.out\n")
-            file.write(f'rm *wfc*\n')
+            if "scf" in self.calculation:
+                file.write(f"mpirun -np {self.ntasks} pw.x -npools {self.npools} < {self.identifier}.scf.in > {self.identifier}.scf.out\n")
+                file.write(f'rm *wfc*\n')
             if "bands" in self.calculation:
                 file.write(f"mpirun -np {self.ntasks} pw.x -npools {self.npools} < {self.identifier}.bands.in > {self.identifier}.bands.out\n")
             if "nscf" in self.calculation:
@@ -556,7 +551,9 @@ class RunScriptHandler():
                         file.write(f"sumpdos.x *\({x[0]}\)* > {self.identifier}.{x[0]}_all.PDOS\n")
                     else:
                         file.write(f"sumpdos.x *\({x[0]}\)*\({x[1]}\) > {self.identifier}.{x[0]}_{x[1]}.PDOS\n")
-            file.write(f'rm *wfc*\n')
+            if "ldos" in self.calculation:
+                file.write(f"mpirun -np {self.ntasks} projwfc.x < {self.identifier}.ldos.in > {self.identifier}.ldos.out\n")
+            # file.write(f'rm *wfc*\n')            
         os.rename(f"{self.identifier}.job", f"./{self.base_folder}/{run_name}/{self.identifier}.job")
 
     def make_runs(self):
